Agents_Systems/JDInterview_Agent/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
streamlit>=1.50.0
openai>=1.0.0
gtts>=2.0.0
